Move all hotels map to map/allHotels
