+++ app/hotels/src/map/allHotels/HotelSwipeItem.js
