+++ app/hotels/src/map/allHotels/HotelSwipeList.js
