+++ app/hotels/src/map/allHotels/MapScreen.js
