+++ app/hotels/src/map/allHotels/MapView.js
