+++ app/hotels/src/map/allHotels/PriceMarker.js
