+++ app/hotels/src/map/allHotels/__tests__/HotelSwipeItem.test.js
@@ -3,7 +3,7 @@
 import * as React from 'react';
 import ShallowRenderer from 'react-test-renderer/shallow';
 
-import { HotelSwipeItem } from './../HotelSwipeItem';
+import { HotelSwipeItem } from '../HotelSwipeItem';
 
 describe('HotelSwipeItem', () => {
   let renderer;
